+++ R/explain.R
@@ -257,20 +257,10 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
     # residual_function not specified
     # try the default
     if (!is.null(predict_function) & model_info$type != "multiclass") {
-      residual_function <- function(model, data, y) {
-        y - predict_function(model, data)
-      }
+      residual_function <- default_residual_function
       verbose_cat("  -> residual function :  difference between y and yhat (",color_codes$yellow_start,"default",color_codes$yellow_end,")\n", verbose = verbose)
     } else if (!is.null(predict_function) & model_info$type == "multiclass") {
-      residual_function <- function(model, data, y) {
-        y_char <- as.character(y)
-        pred <- predict_function(model, data)
-        res <- c()
-        for (i in 1:nrow(pred)) {
-          res[i] <- 1-pred[i, y_char[i]]
-        }
-        res
-      }
+      residual_function <- multiclass_residual_function
       verbose_cat("  -> residual function :  difference between 1 and probability of true class (",color_codes$yellow_start,"default",color_codes$yellow_end,")\n", verbose = verbose)
     }
   } else {
@@ -283,7 +273,7 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
   # if data is specified then we may test residual_function
   residuals <- NULL
   if (!is.null(data) && !is.null(residual_function) && !is.null(y) && (verbose | precalculate)) {
-    residuals <- try(residual_function(model, data, y), silent = TRUE)
+    residuals <- try(residual_function(model, data, y, predict_function), silent = TRUE)
     if (class(residuals)[1] == "try-error") {
       residuals <- NULL
       verbose_cat("  -> residuals         :  the residual_function returns an error when executed (",color_codes$red_start,"WARNING",color_codes$red_end,") \n", verbose = verbose)
@@ -333,6 +323,21 @@ is_y_in_data <- function(data, y) {
 check_if_multilabel <- function(model, predict_function, sample_data) {
   response_sample <- try(predict_function(model, sample_data), silent = TRUE)
   !is.null(dim(response_sample))
+}
+
+default_residual_function <- function(model, data, y, predict_function) {
+  y - predict_function(model, data)
+}
+
+
+multiclass_residual_function <- function(model, data, y, predict_function) {
+  y_char <- as.character(y)
+  pred <- predict_function(model, data)
+  res <- numeric(length(y))
+  for (i in 1:nrow(pred)) {
+    res[i] <- 1-pred[i, y_char[i]]
+  }
+  res
 }
 
 #' @rdname explain