+++ NEWS.md
@@ -1,7 +1,7 @@
 DALEX 1.2.2
 ----------------------------------------------------------------
-* `DALEX` now fully support multiclass cliassification. 
-* `explain()` will use new residual function (1 - true class probability) if multiclass classifiaction is detected.
+* `DALEX` now fully supports multiclass classification. 
+* `explain()` will use new residual function (1 - true class probability) if multiclass classification is detected.
 * `model_performance()` now support measures for multiclass classification.
 
 DALEX 1.2.1